fix type int

+++ ex02/Fixed.hpp
@@ -16,7 +16,7 @@ public:
 	int getRawBits(void) const;
 	void setRawBits(int const raw);
 	float toFloat(void) const;
-	float toInt(void) const;
+	int toInt(void) const;
 
 private:
 	int _nb;
